refactor: use helper funcs to reduce dupe
