+++ plugins/metrics-node/src/definitions/MetricsService.ts
@@ -1,11 +1,7 @@
-import { CounterMetric, UpDownCounterMetric } from '../types';
+import { CounterMetric, UpDownCounterMetric } from '../instruments/counter';
+import { MetricOptions } from '../types';
 
 export interface MetricsService {
   createCounter(name: string, opts?: MetricOptions): CounterMetric;
   createUpDownCounter(name: string, opts?: MetricOptions): UpDownCounterMetric;
-}
-
-export interface MetricOptions {
-  description?: string;
-  labels?: Record<string, string>;
 }
