+++ plugins/metrics-node/src/definitions/index.ts
@@ -1,2 +1,2 @@
-export type { MetricsService, MetricOptions } from './MetricsService';
+export type { MetricsService } from './MetricsService';
 export type { RootMetricsService, MetricsServicePluginOptions } from './RootMetricsService';
