+++ plugins/metrics-node/src/instruments/counter/index.ts
@@ -1,0 +1,4 @@
+export type { CounterMetric, UpDownCounterMetric } from './types';
+
+export { createCounterMetric } from './createCounterMetric';
+export { createUpDownCounterMetric } from './createUpDownCounterMetric';
