+++ plugins/metrics-node/src/types.ts
@@ -1,0 +1,4 @@
+export interface MetricOptions {
+  description?: string;
+  labels?: Record<string, string>;
+}
