--- plugins/metrics-node/src/types/index.ts
@@ -1,1 +1,0 @@
-export type { CounterMetric, UpDownCounterMetric } from './counter';
